Python: executable version! also the evaluation and corresponding plotting is functional now.
Still open:
only the dynamic heuristics and the auctions need to be added to the visitor pattern
however, not all the visitors are very cleanly implemented yet
i also need to figure out how to run the centralized versions

Also:
Maybe having the (initialization, routing, finalization) visitors is not the most sensible way. Instead, instances & carriers could have just one single visitor that in itself combines the necessary parts of initialization+routing+finalization. This visitor could be implemented as a template method