Python
removed the Optimizables' variables that store the init/routing/final visitor & the fields that store whether an instance is init/routed/finalized. These will be moved to a Solution class in the future

fixed some I1 insertion issues, in particular initializing an additional pendulum tour when necessary

on the evaluation plots, algorithms are no longer distinguished by each of init/routing/finalizing visitor but by the overall underlying algorithm (SequentialInsertion, StaticI1, DynamicI1, ...)

on evaluation plots, results are distinguished by the number of available vehicles. This means that centralized instances

 centralized instances are now generated and stored together with the other custom instances (by calling main in instance.py) are can be solved like any other instance

 decoupled the I1 insertion method (incl. c1, c2, ...) from the Instance class. moved it into the routing_visitor.py file