Python:
changed the way that vertices are inserted in tour.py: line 84. instead of pre-filling cost_sequence and schedules with None, they are now filled with 0.

added auction c and changed b; apparently Cluster request selection is much worse than highest marginal insertion cost

added a yml file to be able to recreate the conda environment

added a first attempt of proper logging

added random request selection behavior

created a proper subclass for DynamicSolvers. this helped to reduce code duplication
also now there is a unique solver subclass (of the DynamicSolver) for each auction. This is still cumbersome but works for now

number of requests to be submitted is now an option in utils.py opt dictionary