Python: MASSIVE REFACTORING: Instance and solution are stored separately. Instance stores the bare data (accessible via int indexes) rather than subclasses such as vertices, vehicles, carriers. Designed for Pickup & Delivery Problems. Most parts are still not functional because they have not been adjusted to work with the new way of instances and solutions.
some classes are basically redundant and will soon be deleted: carrier.py, vertex.py, request.py

included a function to read Gansterer & Hartl instances

solutions comprise global solutions (GlobalSolution) and carrier-level (PDPSolution) solutions.

added CheapestInsertion. the previous SequentialCheapestInsertion was different in that it did not compare requests with one another but simply inserted them in the order they came. This CheapestInsertion is basically the only Solver class that is functional atm

got rid of trying to catch InsertionError exceptions outside of the Tour class. There is a feasibility check function to make sure an insertion can be executed.