renamed GlobalSolution to CAHDSolution and CarrierSolution to AHDSolution
ran into issues of infeasibility again. My guess is that it's because of the incremental acceptance phase while during bidding and route_unsbumitted, the routing is done in a static manner

In AuctionD, the route_unsubmitted has been replaced by the dynamic version already

in the GA bundle_generation.py, it is now ensured that the original bundles will always be in the final auction pool. Before, there was a chance that the GA replaced it by (subjectively) better candidates

fixed a mistake in the bundle_valuation.py formula

using rounded distances in the distance matrix from now on. I encountered a local search scenario that infinitely moved from one solution to a second and back.

I have commented the dynamic solvers that run auctions during the acceptance phase. This is material for later. For now, I will stick to running a single auction in the end

added UnequalPreference to the tw_selection.py. This is according to Köhler et al.

cleanup and minor bug fixes