whenever a routing is required, it will follow the dynamic structure of the acceptance phase as closely as possible. This is relevant for (a) the routing of non-submitted requests before the bidding , (b) the bidding phase as well as (c) the final routing. In (a), due to the assigned requests being different than in the acceptance phase, it cannot be guaranteed that a solution will be found.
set up the main.py procedure to ignore failed attempts and to work single core only. I think it makes more sense to implement parallelism e.g. in the bidding (a first attempt made this slower though)

This commit is able to solve many of the instances and on average, the collaboration gain is 2.27% with none of the isolated planning scenarios being worse than its collaborative counterpart.