added: new bundling valuation based on Los et al. (2020). Used for bundle generation, this method considers distance as well as time windows of a bundle
added: a preprocessing phase to bundle generation. The preprocessing depends on the bundling valuation method used and is part of the bundling valuation class, too. This was mainly required to speed up the new bundling valuation method

fixed: a bug with nested loops. In some cases the inner loop iterated over redundant elements. This appeared 2 or 3 times in the code but as far as i can see, this had only increased runtimes but didn't affect the outcome

refactored: move the solver_config list to utils, some renamings to follow the new naming convention for bundle, bundling, bundling_labels, ...