added: using the cprofile and pstats modules for profiling and optimizing the code
optimized: roulette wheel function of the GA was one of the most time consuming functions, optimized the fitness functions (bundle_valuation.py) to return 0 in case of infeasibility or similar. Thus, the weights for random parent selection don't need to be adjusted.

added: random bundle valuation class. Benchmark that creates bundles in the GA randomly by giving each bundle a completely random evaluation. Obviously, no GA with hundreds of generations and candidates is required for this, i might just as well simply return a single random bundling, but doing it this way makes the high level implementation easier

added: a solver_config generator function. makes it a little easier to test and compare different configurations