improved: added a way to automatically set the TIME_MAX for metaheuristics.py to infinity during debugging
improved: parameterized neighborhoods in the parameter generation. Now different sets of neighborhoods can be tested and compared

improved: moved the function to update a solution's solver config into the CAHDSolution class

FixMe: the feasibility check for capacity constraints is wrong!

improved: a bit of cleaning up of the tour_initialization.py as this will be submitted with for the AMO course although it's not in use anymore